requirements.txt
Flask==3.1.2
mysql-connector-python==9.5.0
groq==1.0.0
python-dotenv
gunicorn
